// src/app/models/customerserachpojo.model.ts
export class CustomerSearchPOJO {

	name?: string;
	vorname?: string;
	nachName?: string;
	AntragsNr?: string;
	veranstalter?: string;
	wvStatus?: string;
	swPreisberechnung?: string;
	swinternet?: string;
	statusantrag?: string;
	vorgAnz?: string;
	personalNr?: string;
	statVorgOnline?: string;
	vorgNr?: string;
	antragstArt?: string;
	anzAntragstSteller?: string;
	swVorgStoAntrag?: string;
	swVollStoAntrag?: string;
	vorgNRVon?: string;
	datAntragBegin?: string;
	vorgAnfDat?: string;
	antragsNr?: string;

}